Add APM, dotenv and readme
