+++ examples/national-registration/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "application-name",
+  "name": "mydata-national-registration",
   "version": "0.0.1",
   "scripts": {
     "dev": "nodemon server.js",
@@ -8,6 +8,8 @@
   "dependencies": {
     "@mydata/client": "0.24.1",
     "cookie-parser": "^1.4.4",
+    "dotenv": "^7.0.0",
+    "elastic-apm-node": "^2.8.0",
     "express": "^4.16.4",
     "faker": "^4.1.0",
     "morgan": "^1.9.1",
